src/imageie/MolRuleCoref/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albumentations==1.1.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
easyocr==1.7.2
fastprogress==1.0.3
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.3
grpcio==1.67.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
MolScribe @ git+https://github.com/CrystalEye42/MolScribe.git@250f683f52f5050eb624870ccfd04bccbcaa27e1
multidict==6.1.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
OpenNMT-py==2.2.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
pyclipper==1.3.0.post6
pycocotools==2.0.8
pyonmttok==1.37.1
pyparsing==3.2.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
pytz==2024.2
PyYAML==6.0.2
qudida==0.0.4
rdkit-pypi==2022.9.5
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
smart-open==7.0.5
SmilesPE==0.0.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==0.4.12
tokenizers==0.20.1
torch==1.13.1
torchmetrics==1.5.1
torchtext==0.5.0
torchvision==0.14.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.1
Werkzeug==3.1.1
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
